move OIDC auth specific things to auth-jwt component
